assemblyai/requirements.txt
assemblyai>=0.28.0
python-dotenv>=1.0.0

